refactor(web): move host-context store builders onto the bridges, single store map in HostAdaptor
